requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mongoengine==0.27.0
multidict==6.0.4
pymongo==4.6.1
requests==2.31.0
urllib3==2.1.0
Werkzeug==3.0.1
yarl==1.9.4

python-dotenv~=1.0.0